+++ src/Controller/RespuestaController.php
@@ -27,21 +27,27 @@ class RespuestaController extends AbstractController
         return $this->json($list);
     }
     #[Route('', name: "crear_respuesta", methods: ["POST"])]
-    public function crear_respuesta(EntityManagerInterface $entityManager, Request $request):JsonResponse
+    public function crear_respuesta(EntityManagerInterface $entityManager, Request $request): JsonResponse
     {
-        $json = json_decode($request-> getContent(), true);
+        $json = json_decode($request->getContent(), true);
 
         $nuevarespuesta = new Respuestas();
         $nuevarespuesta->setTexto($json["texto"]);
 
+        // Obtener el tweet basado en el id_tweet
         $tweet = $entityManager->getRepository(Tweets::class)->findOneBy(["id" => $json["id_tweet"]]);
         $nuevarespuesta->setTweet($tweet);
+
+        // Obtener el usuario basado en el id_usuario
+        $usuario = $entityManager->getRepository(Usuario::class)->findOneBy(["id" => $json["id_usuario"]]);
+        $nuevarespuesta->setUsuario($usuario);
 
         $entityManager->persist($nuevarespuesta);
         $entityManager->flush();
 
         return $this->json(['message' => 'Respuesta creada'], Response::HTTP_CREATED);
     }
+
     #[Route('/{id}', name: "editar_respuesta", methods: ["PUT"])]
     public function editar_respuesta(EntityManagerInterface $entityManager, Request $request, Respuestas $respuestas):JsonResponse
     {
@@ -56,13 +62,24 @@ class RespuestaController extends AbstractController
 
         return $this->json(['message' => 'Respuesta modificada'], Response::HTTP_OK);
     }
-    #[Route('/{id}', name: "delete_by_id", methods: ["DELETE"])]
-    public function deleteById_respuesta(EntityManagerInterface $entityManager, Respuestas $respuestas):JsonResponse
+
+
+    #[Route('/{id}', name: "respuesta_delete_by_id", methods: ["DELETE"])]
+    public function deleteById_respuesta(EntityManagerInterface $entityManager, int $id): JsonResponse
     {
+        // Buscar el tweet en la base de datos
+        $respuestas = $entityManager->getRepository(Respuestas::class)->find($id);
+
+        // Verificar si se encontró el tweet
+        if (!$respuestas) {
+            return $this->json(['message' => 'No se encontró el tweet con el ID proporcionado'], Response::HTTP_NOT_FOUND);
+        }
+
+        // Eliminar el tweet
         $entityManager->remove($respuestas);
         $entityManager->flush();
 
-        return $this->json(['message' => 'Respuesta eliminada'], Response::HTTP_OK);
-
+        return $this->json(['message' => 'Tweet eliminado correctamente'], Response::HTTP_OK);
     }
+
 }
